* Rework analyse_compute_blobs_masks.py  - Rename to analyse_blobs_masks.py  - Optimize common variables instantiations  - Optimize multiprocessing using Process and Queue.
